requirements.txt
django
djangorestframework
mysqlclient
Graphene-Django